\name{addLegend}
\alias{addLegend}
\title{Add Legend to a Plot Using Relative (0:1) Coordinates}

\description{
  Places a legend in a plot using relative (0:1) coordinates.
}

\usage{
addLegend(x, y, ...) 
}

\arguments{
  \item{x}{x-axis coordinate in the range (0,1); can step outside}
  \item{y}{y-axis coordinate in the range (0,1); can step outside}
  \item{...}{arguments used by legend, such as \dQuote{lines}, \dQuote{text}, or \dQuote{rectangle}}
}

\seealso{
  \code{\link{addArrows}}
  
  \code{\link{addLabel}}
}

\keyword{iplot}
