\name{clearAll}
\alias{clearAll}
\title{Remove all R Objects From the Global Environment}

\description{
	Generic function to clear .RData in R
}

\usage{clearAll(hidden=TRUE, verbose=TRUE)}

\arguments{
	\item{hidden}{if \code{TRUE}, remove variables that start with a dot (.)}
	\item{verbose}{if \code{TRUE}, report all removed items}
}

\keyword{methods}
