\name{clearWinVal}
\alias{clearWinVal}
\title{Remove All Widget Variables}

\description{
  Removes all global variables that share a name in common with any widget variable name as defined in 
  PBS.win\$vars. Use this with caution.
}

\usage{clearWinVal()}

\seealso{
  \code{\link{getWinVal}}
}

\keyword{methods}
