\name{compareLists}
\alias{compareLists}
\title{Compares two non-recursive lists}

\description{
  Displays any differences between two non-recursive lists.
}

\usage{compareLists(a,b,verbose=FALSE)}

\arguments{
  \item{a}{list a}
  \item{b}{list b}
  \item{verbose}{be verbose}
}

\keyword{methods}
\keyword{list}
