\name{dot-PBSmodEnv}
\docType{data}
\alias{.PBSmodEnv}
\title{PBSmodelling Environment}
\description{
  An environment set aside for PBSmodelling.
}
\usage{.PBSmodEnv}
\format{
A new environment with a \code{.GlobalEnv} parent.
}
\source{
  Generated by a call to the base function \code{new.env()}.
}
\details{
  The environment is created in \code{'zzz.r'} and is used by
  \code{PBSmodelling} functions \code{'lisp'}, \code{'tget'}, 
  \code{'tput'}, \code{'tprint'}, and \code{'tcall'}.
}
\seealso{
In \pkg{PBSmodelling}:\cr
  \code{\link[PBSmodelling]{lisp}},
  \code{\link[PBSmodelling]{tget}},
  \code{\link[PBSmodelling]{packList}},
}
\examples{
lisp(.PBSmodEnv)
}
\keyword{environment}

