\name{getWinFun}
\alias{getWinFun}
\title{Retrieve Names of Functions Referenced in a Window}

\description{
  Get a vector of all function names referenced by a window.
}
\usage{
getWinFun(winName)
}
\arguments{
  \item{winName}{name of window, to retrieve its function list}
}
\value{
  A vector of function names referenced by a window.
}
\author{
  Alex Couture-Beil, Vancouver Island University, Nanaimo BC
}
\keyword{methods}
