\name{isWhat}
\alias{isWhat}
\title{Identify an Object and Print Information}

\description{
  Identify an object by \code{class}, \code{mode}, \code{typeof},
  and \code{attributes}.
}
\usage{
isWhat(x)
}
\arguments{
	\item{x}{an R object}
}
\value{
  No value is returned. The function prints the object's 
  characteristics on the command line.
}
\author{
  Jon Schnute, Pacific Biological Station, Nanaimo BC
}
\keyword{utilities}
