\name{pause}
\alias{pause}
\title{Pause Between Graphics Displays or Other Calculations}
\description{
  Pause, typically between graphics displays. Useful for demo purposes.
}
\usage{
pause(s = "Press <Enter> to continue")
}
\arguments{
  \item{s}{text issued on the command line when \code{pause} is invoked.}
}
\author{
  Jon Schnute, Pacific Biological Station, Fisheries and Oceans Canada, Nanaimo BC
}
\keyword{utilities}
