\name{PBSmodelling}
\docType{package}
\alias{PBSmodelling}
\alias{PBSmodelling-package}
\title{PBS Modelling}

\description{
  PBS Modelling provides software to facilitate the design,
  testing, and operation of computer models. It focuses particularly on
  tools that make it easy to construct and edit a customized graphical
  user interface (GUI). Although it depends heavily on the R interface
  to the Tcl/Tk package, a user does not need to know Tcl/Tk. 
  
  The package contains examples that illustrate models built with other R
  packages, including PBS Mapping, odesolve, and BRugs. It also serves
  as a convenient prototype for building new R packages, along with
  instructions and batch files to facilitate that process. 
  
  The root library directory of PBSmodelling includes a complete user guide
  \code{PBSmodelling-UG.pdf}. To use this package effectively, please consult
  the guide.
  
  PBS Modelling comes packaged with several examples which can be accessed
  with the \code{runExamples()} function.
}

\keyword{package}
