\name{runExamples}
\alias{runExamples}
\title{Display PBS Modelling Examples}

\description{
  Displays an interactive demo GUI to display PBS Modelling examples.

  The example source files can be found in the PBSmodelling/examples directory
  located in the R library.
}
\usage{runExamples()}

\details{
  Some examples make use of external packages which must be installed in order to work correctly.

  LinReg, MarkRec, and CCA Require the Brugs package.

  FishRes requires the odesolve package.

  FishTows requires the PBSmapping package.
}

\keyword{utilities}
