\name{setPBSoptions}
\alias{setPBSoptions}
\title{Set A User Option}

\description{
	Some options may be set by the user.
}

\usage{
setPBSoptions(option, value)
}

\arguments{
	\item{option}{name of the option to set.}
	\item{value}{new value to assign this option.}
}
\seealso{
	\code{\link{getPBSoptions}}
}
\keyword{methods}
