\name{showArgs}
\alias{showArgs}
\title{Display Expected Widget Arguments}

\description{
  Displays the order and default values of widget arguments. The list can be shortened by specifying
  a single widget name. Otherwise all widgets are displayed.
}

\usage{
showArgs(widget="")
}

\arguments{
  \item{widget}{Only displays information about this one widget}
}

\keyword{utilities}
