\name{testLty}
\alias{testLty}
\title{Display Line Types}

\description{
  Displays line types available 
}

\usage{
testLty(newframe = TRUE) 
}

\arguments{
  \item{newframe}{if true, create a new blank frame, otherwise overlay current frame}
}

\keyword{utilities}
\keyword{color}
