\name{vbdata}
\docType{data}
\alias{vbdata}
\title{Dataset: Length-at-Age Data for a von Bertalanffy Curve}

\description{
	Lengths-at-age for freshwater mussels (\emph{Anodonta kennerlyi}).
}
\usage{data(vbdata)}

\format{
	A data frame with 16 rows and 2 columns \code{c("age","len")}.
}
\details{
	Data for demonstartion of the von Bertalanffy model used in the 
	\code{calcMin} example.
}
\source{
	Mittertreiner, A., and Schnute, J. 1985. Simplex: a manual and software package 
	for easy nonlinear parameter estimation and interpretation in fishery research. 
	Canadian Technical Report of Fisheries and Aquatic Sciences 1384: xi + 90 p. 
}
\keyword{datasets}
