\name{view}
\alias{view}
\title{Display First n Rows of an Object}

\description{
  Views first n rows of a data.frame or matrix or first n elements of a vector or list. All other objects are simply reflected.
}
\usage{view(obj, n = 5)}

\arguments{
  \item{obj}{Object to view}
  \item{n}{First n elements of the obj to view}
}

\keyword{print}
