\name{clearWinVal}
\alias{clearWinVal}
\title{Remove all Current Widget Variables}

\description{
	Remove all global variables that share a name in common with any widget variable name 
	defined in \code{names(getWinVal())}. Use this function with caution.
}

\usage{clearWinVal()}

\seealso{
	\code{\link{getWinVal}}
}

\keyword{methods}
