\name{dot-win.funs}
%- Also NEED an '\alias' for EACH other topic documented here.
\alias{.win.chFile}
\alias{.win.chTest}
\alias{.win.closeALL}
\alias{.win.closeChoice}
\alias{.win.closeSDE}
\alias{.win.makeChoice}
\alias{.win.restoreCWD}
\alias{.win.runExHelperQuit}
\alias{.win.tcall}
\alias{.win.tget}
\alias{.win.tprint}
\title{
  GUI Windows Wrappers
}
\description{
  Wrapper functions to run existing or temporary functions from a GUI's function call.
}
\usage{
.win.chFile()
.win.chTest()
.win.closeALL()
.win.closeChoice()
.win.closeSDE()
.win.makeChoice()
.win.restoreCWD()
.win.runExHelperQuit()
.win.tcall()
.win.tget()
.win.tprint()
}
%- maybe also 'usage' for other objects documented here.
\details{
  PBSmodelling Windows (Graphical User Interfaces or GUIs) call
  wrapper functions that call non-GUI functions in the package.
}
\author{
  \href{mailto:rowan.haigh@dfo-mpo.gc.ca}{Rowan Haigh}, Program Head -- Offshore Rockfish\cr
  Pacific Biological Station (PBS), Fisheries & Oceans Canada (DFO), Nanaimo BC\cr
  \emph{locus opus}: Offsite, Vancouver BC\cr
  Last modified \code{Rd: 2023-10-25}
}
%% ~Make other sections like Warning with \section{Warning }{....} ~
\seealso{
In \pkg{PBSmodelling}:\cr
  \code{\link[PBSmodelling]{getWinAct}},
  \code{\link[PBSmodelling]{getWinVal}},
  \code{\link[PBSmodelling]{runExamples}},
  \code{\link[PBSmodelling]{tget}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
\keyword{interface}
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
\concept{supportFuns}
% \concept{ ~cpt2 }
% Use only one concept per line.
