\name{PBSmodelling}
\docType{package}
\alias{PBSmodelling}
\alias{PBSmodelling-package}
\title{PBS Modelling}

\description{
	\emph{PBS Modelling} provides software to facilitate the design, 
	testing, and operation of computer models. It focuses particularly on 
	tools that make it easy to construct and edit a customized graphical 
	user interface (GUI). Although it depends heavily on the R interface 
	to the \code{Tcl/Tk} package, a user does not need to know Tcl/Tk. 

	\code{PBSmodelling} contains examples that illustrate models built uisng 
	other R packages, including \code{PBSmapping}, \code{odesolve}, and 
	\code{BRugs}. It also serves as a convenient prototype for building new R 
	packages, along with instructions and batch files to facilitate that process.

	The root library directory of \code{PBSmodelling} includes a complete 
	user guide \dQuote{PBSmodelling-UG.pdf}. To use this package effectively, 
	please consult the guide.

	\emph{PBS Modelling} comes packaged with several examples that can be 
	accessed using the function \code{runExamples()}.
}

\keyword{package}
