\name{PBSmodelling}
\docType{package}
\alias{PBSmodelling}
\alias{PBSmodelling-package}
\title{PBS Modelling}

\description{
  \emph{PBS Modelling} provides software to facilitate the design, 
  testing, and operation of computer models. It focuses particularly on 
  tools that make it easy to construct and edit a customized graphical 
  user interface (GUI). Although it depends heavily on the R interface 
  to the \code{Tcl/Tk} package, a user does not need to know Tcl/Tk. 

  \code{PBSmodelling} contains examples that illustrate models built using 
  other R packages, including \code{PBSmapping}, \code{odesolve}, 
  \code{PBSddesolve}, and \code{BRugs}. 
  It also serves as a convenient prototype for building new R packages, 
  along with instructions and batch files to facilitate that process.

  The R directory \code{.../library/PBSmodelling/doc} includes a complete 
  user guide \sQuote{PBSmodelling-UG.pdf}. To use this package effectively, 
  please consult the guide.

  \emph{PBS Modelling} comes packaged with interesting examples accessed 
  through the function \code{runExamples()}.
  Additionally, users can view \emph{PBS Modelling} widgets through the 
  function \code{testWidgets()}. 
  More generally, a user can run any available demos in his/her locally 
  installed packages through the function \code{runDemos()}.
}

\keyword{package}
