\name{resetGraph}
\alias{resetGraph}
\title{Reset par Values for a Plot}

\description{
	Reset \code{par()} to default values to ensure that a new plot 
	utilizes a full figure region. This function helps manage the device 
	surface, especially after previous plotting has altered it.
}

\usage{resetGraph()}
\author{
  Jon Schnute, Pacific Biological Station, Nanaimo BC
}
\seealso{
	\code{\link{resetGraph}}
}

\keyword{device}
