\name{showArgs}
\alias{showArgs}
\title{Display Expected Widget Arguments}

\description{
	Display the order and default values of \bold{all} widget arguments. 
	The list can be shortened by specifying a single widget name.
}

\usage{
showArgs(widget="")
}

\arguments{
	\item{widget}{If specified, information about this one widget only is displayed. 
		The default displays information about all widgets.}
}
\value{
	A text stream to the R console. Cannot be directed to a file or other device.
}
\keyword{utilities}
