\name{testLty}
\alias{testLty}
\title{Display Line Types Available}
\description{
  Display line types available.
}
\usage{
testLty(newframe=TRUE, n=1:18, ...)
}
\arguments{
  \item{newframe}{if \code{TRUE}, create a new blank frame, otherwise overlay current frame.}
  \item{n}{vector of line type numbers.}
  \item{...}{additional arguments for function \code{lines}.}
}
\author{
  Rowan Haigh, Pacific Biological Station, Fisheries and Oceans Canada, Nanaimo BC
}
\note{
  Quick representation of line types for reference purposes.
}
\seealso{
  \code{\link[PBSmodelling]{testLwd}}, \code{\link[PBSmodelling]{testPch}}, \code{\link[PBSmodelling]{testCol}}
}
\keyword{utilities}
\keyword{color}

