\name{testLty}
\alias{testLty}
\title{Display Line Types Available}

\description{
	Display line types available.
}

\usage{
testLty(newframe = TRUE)
}

\arguments{
	\item{newframe}{if \code{TRUE}, create a new blank frame, otherwise overlay current frame.}
}
\note{
	Quick representation of first 20 line types for reference purposes.
}
\keyword{utilities}
\keyword{color}
