\name{vbdata}
\docType{data}
\alias{vbdata}
\title{Data: Lengths-at-Age for von Bertalanffy Curve}

\description{
  Lengths-at-age for freshwater mussels (\emph{Anodonta kennerlyi}).
}
\usage{data(vbdata)}

\format{
  A data frame with 16 rows and 2 columns \code{c("age","len")}.
}
\details{
  Data for demonstration of the von Bertalanffy model used in the 
  \code{\link{calcMin}} example.
}
\source{
  Fisheries and Oceans Canada - Mittertreiner and Schnute (1985)
}
\references{
  Mittertreiner, A. and Schnute, J. (1985) Simplex: a manual and software package 
  for easy nonlinear parameter estimation and interpretation in fishery research. 
  \emph{Canadian Technical Report of Fisheries and Aquatic Sciences} \bold{1384}, xi + 90 pp. 
}
\keyword{datasets}
