\name{view}
\alias{view}
\title{Display First n Rows of an Object}

\description{
	View the first \code{n} rows of a data frame or matrix or the first 
	\code{n} elements of a vector or list. All other objects are simply reflected.
}
\usage{view(obj, n = 5)}

\arguments{
	\item{obj}{object to view.}
	\item{n}{first \code{n} rows (matrix/data frame) or elements (vector/list) of \code{obj} to view.}
}
\keyword{print}
